// generated/src/app/api-models/cap-model-select.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapModelSelect {
  capModelId: string;
  seqCapModelSel: number;
  columnName: string;
  columnOccurrence: number;
  operator: string;
  fromValue: string;
  thruValue: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  columnType: string;
}